make new FastPSIModel for specialized code, move datatype helper methods to utils module
